requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
oracledb>=2.0.0
reportlab>=4.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

